requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   google-genai
    #   httpx
cachetools==6.2.4
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
distro==1.9.0
    # via google-genai
google-auth==2.45.0
    # via google-genai
google-genai==1.56.0
    # via tokopedia-fashion-bot (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via tokopedia-fashion-bot (pyproject.toml)
python-telegram-bot==22.5
    # via tokopedia-fashion-bot (pyproject.toml)
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via google-genai
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.2
    # via requests
websockets==15.0.1
    # via google-genai
